# Optimize validator assignment lookups with hybrid O(1) approach for large requests