--- a/changelog/tt_opt-val-lookup.md
+++ b/changelog/tt_opt-val-lookup.md
@@ -1,0 +1,3 @@
+### Changed
+
+- Beacon api optimize validator lookup for large batch request size.